// src/app/modules/pages/my-recipes/models/my-recipes.interface.ts
import { BehaviorSubject } from "rxjs";

export interface RecipesResponseInterface {
  name: string;
  image: string;
  recipes: string;
  userId: string;
}

export interface ColsData {
  field: string;
  header: string;
  type: string;
  userId?: string;
}

export interface GetRecipesModel {
  data: BehaviorSubject<RecipesResponseInterface[]> | RecipesResponseInterface[];
  count: number;
}